--- common/src/test/java/com/microsoft/identity/common/internal/cache/ClientActiveBrokerCacheTest.kt
@@ -1,66 +1,0 @@
-//  Copyright (c) Microsoft Corporation.
-//  All rights reserved.
-//
-//  This code is licensed under the MIT License.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files(the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions :
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-package com.microsoft.identity.common.internal.cache
-
-import com.microsoft.identity.common.java.util.ported.InMemoryStorage
-import kotlinx.coroutines.sync.Mutex
-import org.junit.Assert
-import org.junit.Test
-import java.util.concurrent.TimeUnit
-
-class ClientActiveBrokerCacheTest {
-
-    @Test
-    fun testSetSkipToAccountManager(){
-        val cache = ClientActiveBrokerCache(InMemoryStorage(), Mutex())
-        Assert.assertNull(cache.getCachedActiveBroker())
-
-        Assert.assertFalse(cache.shouldUseAccountManager())
-
-        cache.setShouldUseAccountManagerForTheNextMilliseconds(
-            TimeUnit.SECONDS.toMillis(2)
-        )
-        Assert.assertTrue(cache.shouldUseAccountManager())
-
-        Thread.sleep(TimeUnit.SECONDS.toMillis(2))
-
-        Assert.assertFalse(cache.shouldUseAccountManager())
-    }
-
-    @Test
-    fun testSetSkipToAccountManager_ClearValue(){
-        val cache = ClientActiveBrokerCache(InMemoryStorage(), Mutex())
-        Assert.assertNull(cache.getCachedActiveBroker())
-
-        Assert.assertFalse(cache.shouldUseAccountManager())
-
-        cache.setShouldUseAccountManagerForTheNextMilliseconds(
-            TimeUnit.SECONDS.toMillis(2)
-        )
-        Assert.assertTrue(cache.shouldUseAccountManager())
-
-        cache.clearCachedActiveBroker()
-
-        Assert.assertFalse(cache.shouldUseAccountManager())
-    }
-}
